arne

+++ launch_chffrplus.sh
@@ -6,10 +6,13 @@ fi
 
 function launch {
   # apply update
-  if [ "$(git rev-parse HEAD)" != "$(git rev-parse @{u})" ]; then
-     git reset --hard @{u} &&
-     git clean -xdf &&
-     exec "${BASH_SOURCE[0]}"
+  file="/data/no_ota_updates"
+  if ! [ -f "$file" ]; then
+    if [ "$(git rev-parse HEAD)" != "$(git rev-parse @{u})" ]; then
+       git reset --hard @{u} &&
+       git clean -xdf &&
+       exec "${BASH_SOURCE[0]}"
+    fi
   fi
 
   # no cpu rationing for now
@@ -19,7 +22,12 @@ function launch {
   echo 0-3 > /dev/cpuset/foreground/cpus
   echo 0-3 > /dev/cpuset/android/cpus
 
+  # handle pythonpath
+  ln -s /data/openpilot /data/pythonpath
   export PYTHONPATH="$PWD"
+
+  # check to see if custom APK is installed
+  python ./checkCustomAPK.py
 
   # start manager
   cd selfdrive
